Fixes build error
